requirements.txt
fastapi
uvicorn
psutil
jinja2
psycopg2-binary
apscheduler
